Work done on the display of a show's reviews, mainly improved UI, but added feature to click on review and view text. I had changed my api call for retrieving the reviews a while back and realised today that I broke it, that is now fixed. Also some work done on displaying pop up dialog when filtering by location, again the functionality was already there but UI improved. Have also realised that the feature I had for checking if a new show instance overlaps with an exising one no longer works - the code I had is now only there in part

+++ CompSci Project/app/src/main/java/com/example/theatreticketsapp/Show.java
@@ -13,7 +13,7 @@ public class Show implements Parcelable {
     private int monMat, monEve, tueMat, tueEve, wedMat, wedEve, thuMat, thuEve, friMat, friEve, satMat, satEve, sunMat, sunEve;
     private int pricePbA, pricePbB, pricePbC, pricePbD, numTicketsPbA, numTicketsPbB, numTicketsPbC, numTicketsPbD;
     private int id;
-    private int rating;
+    private double rating;
     private int runningTime;
     private Venue venue;
     private float userDistanceFromVenue;
@@ -72,7 +72,7 @@ public class Show implements Parcelable {
     public static Comparator<Show> compareRating = new Comparator<Show>() {
         @Override
         public int compare(Show s1, Show s2) {
-            return (Integer.compare(s2.getRating(), s1.getRating()));
+            return (Double.compare(s2.getRating(), s1.getRating()));
         }
     };
 
@@ -81,6 +81,7 @@ public class Show implements Parcelable {
 
 
     //GETTERS AND SETTERS:
+
 
     public void setShowName(String s){
         this.venueName = s;
@@ -228,11 +229,11 @@ public class Show implements Parcelable {
         return numTicketsPbD;
     }
 
-    public void setRating(int avgRating){
+    public void setRating(double avgRating){
         this.rating = avgRating;
     }
 
-    public int getRating(){
+    public double getRating(){
         return this.rating;
     }
 
@@ -285,7 +286,7 @@ public class Show implements Parcelable {
         numTicketsPbC = in.readInt();
         numTicketsPbD = in.readInt();
 
-        rating = in.readInt();
+        rating = in.readDouble();
         runningTime = in.readInt();
 
         venue = in.readParcelable(Venue.class.getClassLoader());
@@ -346,7 +347,7 @@ public class Show implements Parcelable {
         dest.writeInt(numTicketsPbC);
         dest.writeInt(numTicketsPbD);
 
-        dest.writeInt(rating);
+        dest.writeDouble(rating);
         dest.writeInt(runningTime);
 
         dest.writeParcelable(venue, flags);
